#7: moved fakes/ to src/ and tests/fakes/ to tests/src/
